# [nexmark] Use Runtime to support multiple cores.