--- a/Cargo.toml
+++ b/Cargo.toml
@@ -53,6 +53,7 @@ clap = { version = "3.2.8", features = ["derive", "env"] }
 reqwest = { version = "0.11.11", features = ["blocking"] }
 ascii_table = "4.0.2"
 num-format = "0.4.0"
+pbr = "1.0.4"
 
 [target.'cfg(unix)'.dev-dependencies]
 libc = "0.2.127"