--- a/Cargo.toml
+++ b/Cargo.toml
@@ -53,7 +53,10 @@ clap = { version = "3.2.8", features = ["derive", "env"] }
 reqwest = { version = "0.11.11", features = ["blocking"] }
 ascii_table = "4.0.2"
 num-format = "0.4.0"
-pbr = "1.0.4"
+# The latest release of pbr (1.0.4) depends on an old version of time (0.1.44) with
+# a vulnerability. The dependency has been removed but a new version is yet
+# to be released https://github.com/a8m/pb/issues/113
+pbr = { git = "https://github.com/a8m/pb", rev = "09a8e592c1bb0aa1d6215e35c5c8b49b7a5ad6bd"}
 
 [target.'cfg(unix)'.dev-dependencies]
 libc = "0.2.127"