--- a/benches/nexmark/main.rs
+++ b/benches/nexmark/main.rs
@@ -8,7 +8,7 @@ use libc::{getrusage, rusage, timeval, RUSAGE_THREAD};
 use std::{
     io::Error,
     mem::MaybeUninit,
-    sync::{mpsc, mpsc::TryRecvError},
+    sync::mpsc,
     thread,
     time::{Duration, Instant},
 };
@@ -25,7 +25,7 @@ use dbsp::{
         NexmarkSource,
     },
     trace::{ord::OrdZSet, BatchReader},
-    Circuit,
+    Circuit, Runtime,
 };
 use num_format::{Locale, ToFormattedString};
 use rand::prelude::ThreadRng;
@@ -35,47 +35,23 @@ use rand::prelude::ThreadRng;
 // haven't yet found a way to import from a `lib.rs` in the same directory as
 // the benchmark's `main.rs`)
 
-/// Returns a closure for a circuit with the nexmark source that sets
-/// `max_events_reached` once no more data is available.
+/// Returns a closure for a circuit with the nexmark source that returns
+/// the input handle.
 macro_rules! nexmark_circuit {
-    ( $q:expr, $generator_config:expr, $output_fixedpoint_tx:expr ) => {
+    ( $q:expr ) => {
         |circuit: &mut Circuit<()>| {
-            // We need the source to signal when it has reached its fixedpoint.
-            let (fixedpoint_tx, fixedpoint_rx) = mpsc::channel();
-            let mut num_events_generated = 0;
+            let (stream, input_handle) = circuit.add_input_zset::<Event, isize>();
 
-            let source = NexmarkSource::<ThreadRng, isize, OrdZSet<Event, isize>>::new(
-                $generator_config,
-                fixedpoint_tx,
-            );
-            let input = circuit.add_source(source);
-
-            let output = $q(input);
+            let output = $q(stream);
 
             output.inspect(move |zs: &OrdZSet<_, _>| {
-                // Turns out we can't get an exact count of events by accumulating the lengths
-                // of the `OrdZSet` since duplicate bids are not that difficult to
-                // produce in the generator (0-3 per 1000), which get merged to a
-                // single Item in the `OrdZSet` with an adjusted weight. Nor does
-                // the number of output events even correspond with the input events
-                // necessarily. Nor can we rely on empty sets as an indicator that
-                // it's finished (since they'll be returned in queries that don't
-                // emit data often). So, for simplicity, the source sends a message
-                // on it's fixedpoint_sync channel when it reaches its fixed point, which we can
-                // read here and forward to the test harness.
-                match fixedpoint_rx.try_recv() {
-                    Ok(num_events) => {
-                        num_events_generated = num_events;
-                    }
-                    Err(TryRecvError::Empty) => (),
-                    _ => panic!("unexpected result from fixedpoint_sync channel"),
-                };
-                // The source may reach its fixed point for inputs before they have been
-                // processed.
-                if num_events_generated > 0 && zs.len() == 0 {
-                    $output_fixedpoint_tx.send(num_events_generated).unwrap();
-                }
+                // Currently using the print below not only to get an idea of the batch
+                // size but also to see how much time is spent creating the zsets from
+                // the input.
+                println!("zs.len() = {}", zs.len());
             });
+
+            input_handle
         }
     };
 }
@@ -91,28 +67,23 @@ struct NexmarkResult {
 }
 
 macro_rules! run_query {
-    ( $q_name:expr, $q:expr, $generator_config:expr, $max_events:expr, $result_tx:expr ) => {{
-        // Until we have the I/O API to control the running of circuits,
-        // use a channel to signal when the test is finished (all events emitted
-        // by the generator), sending back the number of events processed.
-        let (fixedpoint_tx, fixedpoint_rx): (mpsc::Sender<u64>, mpsc::Receiver<u64>) =
-            mpsc::channel();
+    ( $q_name:expr, $q:expr, $generator_config:expr, $result_tx:expr ) => {{
+        let circuit_closure = nexmark_circuit!($q);
 
-        let circuit = nexmark_circuit!($q, $generator_config, fixedpoint_tx);
+        let num_cores = $generator_config.nexmark_config.cpu_cores;
+        let (mut dbsp, mut input_handle) =
+            Runtime::init_circuit(num_cores, circuit_closure).unwrap();
 
-        let root = Circuit::build(circuit).unwrap().0;
+        let source =
+            NexmarkSource::<ThreadRng, isize, OrdZSet<Event, isize>>::new($generator_config);
 
-        let num_events_generated;
+        let mut num_events_generated: u64 = 0;
         let start = Instant::now();
-        loop {
-            match fixedpoint_rx.try_recv() {
-                Ok(num_events) => {
-                    num_events_generated = num_events;
-                    break;
-                }
-                Err(TryRecvError::Empty) => root.step().unwrap(),
-                _ => panic!("unexpected result from fixedpoint sync channel"),
-            }
+
+        for mut batch in source {
+            num_events_generated += batch.len() as u64;
+            input_handle.append(&mut batch);
+            dbsp.step().unwrap();
         }
 
         let (usr_cpu, sys_cpu, max_rss) = unsafe { rusage_thread() };
@@ -146,7 +117,7 @@ macro_rules! run_queries {
             let thread_result_tx = result_tx.clone();
             let thread_generator_config = $generator_config.clone();
             thread::spawn(move || {
-                run_query!($q_name, $q, thread_generator_config, $max_events, thread_result_tx);
+                run_query!($q_name, $q, thread_generator_config, thread_result_tx);
             });
             // Wait for the thread to finish then collect the result.
             results.push(result_rx.recv().unwrap());
@@ -158,16 +129,16 @@ macro_rules! run_queries {
 
 fn create_ascii_table() -> AsciiTable {
     let mut ascii_table = AsciiTable::default();
+    ascii_table.set_max_width(120);
     ascii_table.column(0).set_header("Query");
     ascii_table.column(1).set_header("#Events");
     ascii_table.column(2).set_header("Cores");
-    ascii_table.column(3).set_header("Elapsed(s)");
-    // Redundant until we use more than one core.
-    // ascii_table.column(4).set_header("Cores * Time(s)");
-    ascii_table.column(4).set_header("Throughput/Cores");
-    ascii_table.column(5).set_header("User CPU(s)");
-    ascii_table.column(6).set_header("System CPU(s)");
-    ascii_table.column(7).set_header("Max RSS(Kb)");
+    ascii_table.column(3).set_header("Time(s)");
+    ascii_table.column(4).set_header("Cores * Time(s)");
+    ascii_table.column(5).set_header("Throughput/Cores");
+    ascii_table.column(6).set_header("User CPU(s)");
+    ascii_table.column(7).set_header("System CPU(s)");
+    ascii_table.column(8).set_header("Max RSS(Kb)");
     ascii_table
 }
 
@@ -192,6 +163,7 @@ fn main() -> Result<()> {
     let nexmark_config = NexmarkConfig::parse();
     let max_events = nexmark_config.max_events;
     let queries_to_run = nexmark_config.query.clone();
+    let cpu_cores = nexmark_config.cpu_cores;
     let generator_config = GeneratorConfig::new(nexmark_config, 0, 0, 0);
 
     let results = run_queries!(
@@ -211,11 +183,12 @@ fn main() -> Result<()> {
         vec![
             r.name,
             format!("{}", r.num_events.to_formatted_string(&Locale::en)),
-            String::from("1"),
+            format!("{}", cpu_cores),
             format!("{0:.3}", r.elapsed.as_secs_f32()),
+            format!("{0:.3}", cpu_cores as f32 * r.elapsed.as_secs_f32()),
             format!(
                 "{0:.3} K/s",
-                r.num_events as f32 / r.elapsed.as_secs_f32() / 1000.0
+                r.num_events as f32 / r.elapsed.as_secs_f32() / cpu_cores as f32 / 1000.0
             ),
             format!("{0:.3}", r.usr_cpu.as_secs_f32()),
             format!("{0:.3}", r.sys_cpu.as_secs_f32()),