@absoludity 2022-08-19
Summary of chat with @ryzhyk 
- Looks like time might be mostly spent injecting the data
- Can check the time of the `step` call vs the `append` call. `append` is the one sharding inputs across 10 threads.
- Related to that, when using 10M events per second, the data is generated and input as a single batch (limited there to 1M records by `max_events`), because the batch only closes when the next generated event is still in the future (and 10M/second means they're never in the future, it seems?)
- May want to add a condition so the batch closes when a max size (100k) is reached, or as soon as DBSP is idle and ready to accept more data.
- Could run generator in separate thread filling a buffered sync channel, that way whenever `dbsp.step()` completes we can call `input_handle.append()` with the content of the channel. That way dbsp always processes the next batch of data when it is idle, and the generator has back-pressure from the buffer size of the sync channel.