@ryzhyk 2022-08-19
Forgot to mention that it should be totally ok to call 'append' even while dbsp is running. Appended data will get consumed by the next 'step' call. So perhaps the right model is for the generator to append generated records in small batches, eg., 1000 records, with a separate thread simply calling 'step' in a loop.