@absoludity 2022-08-21
> Forgot to mention that it should be totally ok to call 'append' even while dbsp is running. Appended data will get consumed by the next 'step' call. So perhaps the right model is for the generator to append generated records in small batches, eg., 1000 records, with a separate thread simply calling 'step' in a loop.

Except that the `CollectionHandle` will apply zero back-pressure, given that the buffers it uses internally are unlimited? So the generator can just OOM the setup by `append`ing data much faster than it can be consumed. I guess we can't limit the buffer used by `CollectionHandle` in some (configurable) way, longer-term, since it would mean it'd be blocking. Even if the `CollectionHandle` enabled users to know how many input(set)s it's currently buffering, the user could decide whether they want to limit the input buffer themselves.