@ryzhyk 2022-08-21
If the generator thread runs concurrently with the thread that calls `step`, then an OOM situation can totally happen, so we still need some form of backpressure.