@ryzhyk 2022-08-21
One option would be to extend the handle API to report the number of buffered records.  The generator would use that to bound the amount of buffered data.