@absoludity 2022-08-22
> One option would be to extend the handle API to report the number of buffered records. The generator would use that to bound the amount of buffered data.

Yeah, that's what I thought too (see above comment), but looking at the code in `input.rs`, I'd need to return the `T.len()` of the `Mailbox<T>`, but don't see how we can do that currently when `T` is generic (and there's not trait for `len` than `Vec` implements). I think it'd require updating so that `Mailbox<T>` stored at `Arc<Mutex<Vec<T>>>`, but I don't know if you need mailbox to be completely generic? (seems from the code that you want that).