@ryzhyk 2022-08-22
Yes, it would take a bit of refactoring, but it's doable.  I am finishing another PR now, but can do `CollectionHandle::len()` next.