@absoludity 2022-08-22
> Yes, it would take a bit of refactoring, but it's doable. I am finishing another PR now, but can do `CollectionHandle::len()` next.

I took a different approach to coordinate how much data is inputted dynamically, ensuring that it continues to increase the input as long as dbsp processes the data more quickly than it is added. With this, we can test with 100M events without memory issues. See the updated description.