@absoludity 2022-08-23
OK, updated a ready for review/landing. See the updated description for the improvements in UX and cpu-usage calculation. I'll try reducing the input times that we chatted about @ryzhyk in a separate PR.